Add files via upload
This code works for only one finger. It is very hard coded; however, it works to change the voltage needed to change a motor as the angle of the finger gets closer to the target angle. It does not fully work, but it will change further to accommodate any n number of fingers. This code is simply modified code of the Get Pot Vals code.